@@ -2,18 +2,18 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.15.8" />
+	  <PackageReference Include="Microsoft.VisualStudio.DiagnosticsHub.BenchmarkDotNetDiagnosers" Version="18.0.36313.1" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\CelSerEngine.Core\CelSerEngine.Core.csproj" />
-    <ProjectReference Include="..\CelSerEngine.Core.IntegrationTests\CelSerEngine.Core.IntegrationTests.csproj" />
   </ItemGroup>
 
 </Project>
