@@ -1,6 +1,6 @@
 ﻿namespace CelSerEngine.Core.Scanners;
 
-internal class PointerData
+internal struct PointerData
 {
     public IntPtr Address { get; set; }
     public StaticData? StaticData { get; set; }
