@@ -1,6 +1,6 @@
 ﻿namespace CelSerEngine.Core.Scanners;
 
-internal class StaticData
+internal struct StaticData
 {
     public int ModuleIndex { get; set; }
     public IntPtr Offset { get; set; }
